aitalkmaster-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
openai==1.72.0
pydantic==2.5.0
PyYAML==6.0.1
ollama
mutagen==1.47.0
requests
